--- src/constants/proto-type-names.constants.js
@@ -1,8 +1,0 @@
-const protoTypeNames = {
-  // GAME_PACKET: "GamePacket",
-  REQUEST: "RequestPacket",
-  RESPONSE: "ResponsePacket",
-  NOTIFICATION: "NotificationPacket",
-};
-
-export default protoTypeNames;
